feat(control-api): recovery-only route gate at the protocol boundary
resolveControlProtocol (extracted to control-protocol.ts) now takes the
serving mode: the handshake discloses it and, while recovery_only,
non-recovery product routes are refused with a typed retryable 503 before
any handler runs. Request-body reading extracted to request-body.ts to
hold the daemon-server complexity ratchet.
