+++ packages/control-api/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./daemon-server.js";
+export * from "./control-protocol.js";
 export * from "./operation-catalog.js";
 export * from "./retention.js";
 export { normalizeExistingProjectRoot } from "./run-start.js";
