+++ packages/control-api/src/operation-catalog.ts
@@ -1,110 +1,13 @@
 import {
   CONTROL_PROTOCOL_MAJOR,
-  ControlHandshakeRequest,
-  ControlHandshakeResponse,
   ControlOperationCatalog,
-  ControlProblem,
   ControlRunState,
   type ControlOperationDescriptor,
 } from "@claudexor/schema";
-import { engineBuildIdentity } from "@claudexor/util";
-import { pathnameDecodes, queryParam, resumeHeader } from "./operation-parameters.js";
+import { queryParam, resumeHeader } from "./operation-parameters.js";
 import { REMOTE_OPERATION_DRAFTS } from "./remote-operation-descriptors.js";
 import type { OperationDraft } from "./operation-draft.js";
 import { OPERATION_SUMMARIES } from "./operation-summaries.js";
-
-export type ControlProtocolBoundary =
-  | { kind: "route"; path: string }
-  | { kind: "response"; status: number; body: unknown; contentType: string };
-
-const protocolProblem = (code: string, message: string, requiredActions: string[] = []) =>
-  ControlProblem.parse({
-    code,
-    message,
-    retryable: false,
-    fieldErrors: {},
-    requiredActions,
-    evidenceRefs: [],
-  });
-
-/** Stateless v2 negotiation boundary; product handlers only see unversioned internal paths. */
-export async function resolveControlProtocol(input: {
-  method: string;
-  requestPath: string;
-  requestedMajor: string | string[] | undefined;
-  readBody: () => Promise<unknown>;
-}): Promise<ControlProtocolBoundary> {
-  if (input.method === "POST" && input.requestPath === "/v2/handshake") {
-    const request = ControlHandshakeRequest.parse(await input.readBody());
-    if (request.protocolMajor !== CONTROL_PROTOCOL_MAJOR) {
-      return {
-        kind: "response",
-        status: 426,
-        contentType: "application/problem+json",
-        body: protocolProblem(
-          "incompatible_protocol_major",
-          `control protocol major ${request.protocolMajor} is incompatible; server requires ${CONTROL_PROTOCOL_MAJOR}`,
-          [`use control protocol major ${CONTROL_PROTOCOL_MAJOR}`],
-        ),
-      };
-    }
-    return {
-      kind: "response",
-      status: 200,
-      contentType: "application/json",
-      body: ControlHandshakeResponse.parse({
-        protocolMajor: CONTROL_PROTOCOL_MAJOR,
-        compatible: true,
-        operationsPath: "/v2/operations",
-        engine: engineBuildIdentity(),
-      }),
-    };
-  }
-  if (!input.requestPath.startsWith("/v2/")) {
-    return {
-      kind: "response",
-      status: 404,
-      contentType: "application/problem+json",
-      body: protocolProblem("route_not_found", "product routes require the /v2 prefix"),
-    };
-  }
-  if (input.requestedMajor !== String(CONTROL_PROTOCOL_MAJOR)) {
-    return {
-      kind: "response",
-      status: 426,
-      contentType: "application/problem+json",
-      body: protocolProblem(
-        "handshake_required",
-        "a successful v2 handshake is required before product calls",
-        ["POST /v2/handshake", `send X-Claudexor-Protocol-Major: ${CONTROL_PROTOCOL_MAJOR}`],
-      ),
-    };
-  }
-  if (input.method === "GET" && input.requestPath === "/v2/operations") {
-    return {
-      kind: "response",
-      status: 200,
-      contentType: "application/json",
-      body: OPERATION_CATALOG,
-    };
-  }
-  // QA-066: malformed percent-encoding in the path is a CLIENT syntax error —
-  // validate the whole encoded pathname decodes ONCE, centrally, before route
-  // dispatch (typed 400, not a per-route URIError into the 500 handler).
-  // Routes still match on the ENCODED path; `%2F`/`%2e%2e` semantics unchanged.
-  if (!pathnameDecodes(input.requestPath)) {
-    return {
-      kind: "response",
-      status: 400,
-      contentType: "application/problem+json",
-      body: protocolProblem(
-        "malformed_request_path",
-        "request path contains malformed percent-encoding",
-      ),
-    };
-  }
-  return { kind: "route", path: input.requestPath.slice(3) };
-}
 
 function descriptor(input: OperationDraft): ControlOperationDescriptor {
   // Resource-family classification (QA-054): an operation is grouped under the
